requirements.txt
torch>=2.1.0
transformers>=4.45.0
peft>=0.13.0
accelerate>=0.33.0
bitsandbytes>=0.43.0
datasets>=2.19.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.66.0
pyyaml>=6.0
requests>=2.31.0
openai>=1.6.0
dashscope>=1.18.0
streamlit>=1.32.0
huggingface_hub>=0.20.2
